+++ client/src/App.css
@@ -1,12 +1,12 @@
 .App {
   text-align: center;
 }
-*{
+/* *{
   box-sizing: border-box;
   margin: 0 auto;
   padding: 0;
   font-family: 'Encode Sans Expanded',
-}
+} */
 
 
 @font-face {
