+++ client/src/components/views/InfoSection/InfoElements.js
@@ -1,9 +1,18 @@
 import styled from 'styled-components';
 
+
 export const InfoContainer = styled.div`
-    
     color: #fff;
-    background: ${({ lightBg }) => (lightBg ? '#f9f9f9' : '#010606')};
+    /* background: ${({ lightBg }) => (lightBg ? '#f9f9f9' : '#010606')}; */
+    /* background: #010606; */
+    background: #f9f9f9;
+
+    * {
+        box-sizing: border-box;
+        margin: 0 auto;
+        padding: 0;
+    }
+
 
     @media screen and (max-width: 768px){
         padding: 100px 0;
@@ -13,11 +22,13 @@ export const InfoContainer = styled.div`
 export const InfoWrapper = styled.div`
     display: grid;
     z-index: 1;
-    height: 860px;
+    /* height: 860px; */
     width: 100%;
     max-width: 1100px;
+    margin: 200px auto;
+    /* margin-top: 200px;
     margin-right: auto;
-    margin-left: auto;
+    margin-left: auto; */
     padding: 0 24px;
     justify-content: center;
 `
@@ -52,7 +63,8 @@ export const TextWrapper = styled.div`
 `
 
 export const TopLine = styled.p`
-    color: #01bf71;
+    /* color: #01bf71; */
+    color: #000;
     font-size: 16px;
     line-height: 16px;
     font-weight: 700;
@@ -69,6 +81,8 @@ export const Heading = styled.h1`
     letter-spacing: 2px;
     font-weight: 600;
     color: ${({ lightText }) => (lightText ? '#f7f8fa' : '#010606')};
+    color: #000;
+    
 
     @media screen and (max-width: 480px){
         font-size: 32px;
@@ -81,12 +95,36 @@ export const Subtitle = styled.p`
     font-size: 18px;
     line-height: 24px;
     color: ${({ darkText }) => (darkText ? '#010606' : '#fff')};
+    color: #000;
+    
 `
 
 export const BtnWrap = styled.div`
     display: flex;
     justify-content: flex-start;
     margin-right: 20px;
+`
+
+export const Buttons = styled.div`
+    border-radius: 50px;
+    background: ${({ primary }) => (primary ? '#01BF71' : '#010606')};
+    white-space: nowrap;
+    padding: ${({ big }) => (big ? '14px 48px' : '12px 30px')};
+    color: ${({ dark }) => (dark ? '#010606' : '#fff')};
+    font-size: ${({ fontBig }) => (fontBig ? '20px' : '16px')};
+    outline: none;
+    border: none;
+    cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    transition: all 0.2s ease-in-out;
+
+    &:hover{
+        transition: all 0.2s ease-in-out;
+        background: ${({ primary }) => (primary ? '#fff' : '#01BF71')};
+        color: ${({ primary }) => (primary ? '#010606' : '#fff')};
+    }
 `
 
 export const ImgWrap = styled.div`
